+++ Markdown Viewer/src/bootstrap.js
@@ -45,6 +45,7 @@ const applicationModuleStages = [
       () => import('./components/controllers/TabSessionController.js'),
       () => import('./components/controllers/DocumentLifecycleController.js'),
       () => import('./components/controllers/EditorLifecycleController.js'),
+      () => import('./components/controllers/PreviewLifecycleController.js'),
       () => import('./components/controllers/MarkdownActionController.js'),
       () => import('./components/controllers/ExportController.js'),
       () => import('./components/controllers/NativeWindowController.js'),